requirements.txt
asgiref==3.6.0
boto3==1.26.60
dj-database-url==1.2.0
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
django-multiselectfield==0.1.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-rest-multiple-models==2.1.3
django-storages==1.13.2
django-push-notifications==3.0.0
django-celery-beat==2.5.0
django-simple-history==3.3.0
drf-yasg==1.21.5
celery==5.2.7
gunicorn==20.1.0
python-decouple==3.7
pytz==2022.7.1
sqlparse==0.4.3
twilio==7.16.3
redis==4.5.4
Pillow==9.5.0
flower==1.2.0
